move make_cog functions to image.tiff
