+++ hyp3lib/depreciated/make_cogs.py
@@ -3,39 +3,9 @@
 import argparse
 import logging
 import os
-import shutil
 import sys
-from glob import glob
-from tempfile import NamedTemporaryFile
 
-from osgeo import gdal
-
-
-def cogify_dir(directory: str, file_pattern: str = '*.tif'):
-    """
-    Convert all found GeoTIFF files to a Cloud Optimized GeoTIFF inplace
-    Args:
-        directory: directory to search through
-        file_pattern: the pattern for finding GeoTIFFs
-    """
-    path_expression = os.path.join(directory, file_pattern)
-    logging.info(f'Converting files to COGs for {path_expression}')
-    for filename in glob(path_expression):
-        cogify_file(filename)
-
-
-def cogify_file(filename: str):
-    """
-    Convert a GeoTIFF to a Cloud Optimized GeoTIFF inplace
-
-    Args:
-        filename: GeoTIFF file to convert
-    """
-    logging.info(f'Converting {filename} to COG')
-    creation_options = ['TILED=YES', 'COMPRESS=DEFLATE']
-    with NamedTemporaryFile() as temp_file:
-        shutil.copy(filename, temp_file.name)
-        gdal.Translate(filename, temp_file.name, format='GTiff', creationOptions=creation_options, noData=0)
+from hyp3lib.image.tiff import cogify_file
 
 
 def main():
